fix: added equals and hashcode for entities

+++ src/main/java/com/back/backend/service/GameService.java
@@ -32,22 +32,22 @@ public class GameService {
 
     public Room createRoom(String name) {
         Room room = new Room();
-        room.setGame(new Game());          // что делать с
+        room.setGame(new Game());
         room.setName(name);
         room.setMaxCount(2);
         room.setCount(0);
         return roomRepository.save(room);
     }
 
-    public Optional<Player> getPerson(int id) {
+    public Optional<Player> getPerson(long id) {
         Player player = new Player();
-        Optional<Player> byId = playerRepository.findById((long) id);
+        Optional<Player> byId = playerRepository.findById(id);
         return byId;
     }
 
-    public  Optional<Room> roomById(int id) {
+    public  Optional<Room> roomById(long id) {
         Room room = new Room();
-        Optional<Room> byId = roomRepository.findById((long) id);
+        Optional<Room> byId = roomRepository.findById(id);
         return byId;
     }
 }
